fix: resolve all backend startup blockers after merge from main
- Remove V4__placeholder.sql (conflicted with V04__create_teacher_tables.sql)
- Remove duplicate CorsConfig.java bean (SecurityConfig already has CORS)
- Strip V03 migration to no-op (tables already created in V02)
- Strip V07 migration to no-op (subjects already created in V06)
- Rename duplicate index idx_assignments_class_group in V11 to idx_assignments_cg
- Add @Component to JwtAuthenticationFilter and JwtRequestAttributeFilter
- Add @Lazy to UserDetailsService injection to break circular dependency
- Fix GradingScaleServiceImpl: remove doubleValue()/BigDecimal.valueOf() mismatches
- Add createDatabaseIfNotExist=true to JDBC URL for zero-config developer setup
- Use env vars DB_USERNAME/DB_PASSWORD with postgres defaults

+++ backend/elmkusoma-core/src/main/java/tz/elmkusoma/grading/service/impl/GradingScaleServiceImpl.java
@@ -30,8 +30,8 @@ public class GradingScaleServiceImpl implements GradingScaleService {
                 .name(request.getName())
                 .description(request.getDescription())
                 .scaleType(GradingScale.ScaleType.valueOf(request.getScaleType()))
-                .minValue(request.getMinValue().doubleValue())
-                .maxValue(request.getMaxValue().doubleValue())
+                .minValue(request.getMinValue())
+                .maxValue(request.getMaxValue())
                 .isDefault(request.getIsDefault())
                 .build();
 
@@ -66,8 +66,8 @@ public class GradingScaleServiceImpl implements GradingScaleService {
         scale.setName(request.getName());
         scale.setDescription(request.getDescription());
         scale.setScaleType(GradingScale.ScaleType.valueOf(request.getScaleType()));
-        scale.setMinValue(request.getMinValue().doubleValue());
-        scale.setMaxValue(request.getMaxValue().doubleValue());
+        scale.setMinValue(request.getMinValue());
+        scale.setMaxValue(request.getMaxValue());
         scale.setIsDefault(request.getIsDefault());
 
         GradingScale saved = gradingScaleRepository.save(scale);
@@ -99,8 +99,8 @@ public class GradingScaleServiceImpl implements GradingScaleService {
                 .name(scale.getName())
                 .description(scale.getDescription())
                 .scaleType(scale.getScaleType().name())
-                .minValue(BigDecimal.valueOf(scale.getMinValue()))
-                .maxValue(BigDecimal.valueOf(scale.getMaxValue()))
+                .minValue(scale.getMinValue())
+                .maxValue(scale.getMaxValue())
                 .isDefault(scale.getIsDefault())
                 .isActive(scale.getIsActive())
                 .createdAt(scale.getCreatedAt())
